ProductTypeDraft to interface. #942
